category default limit

+++ packages/Sarga/Admin/src/Database/Migrations/2022_03_09_212834_add_limit_to_categories_table.php
@@ -14,7 +14,7 @@ class AddLimitToCategoriesTable extends Migration
     public function up()
     {
         Schema::table('categories', function (Blueprint $table) {
-            $table->decimal('product_limit')->unsigned()->default(500);
+            $table->integer('product_limit')->unsigned()->default(500);
 
         });
     }
